refactor: move LRC parsing functions to core/utils
Move parse_lrc, ms_to_ts, parse_ts_str, and _ts_to_ms from
lyrics_editor_widget to core/utils for reuse across modules.
